VoiceToText functionality extended
List of ignored words is now dynamic and accessible through the inspector. The inspector will also show the last voice segment as a read-only field for testing purposes.

OnSegmentFinish will now also filter anything in round brackets as well, I didn't know it indicates environmental sounds (like running water, claps etc.) in whisper apparently.

Word sorting function now returns a sorted dictionary instead of an array. The reason is to preserve the information about the number of times a word have been used, in case we want to display it in the interface (which we probably do). Also the function itself is now public as it should be. As well as the text string, in case we'll need raw unfiltered transcription of the entire session for some reason.

A public Reset functionality was added to Reset the VoiceToText state in between sessions, as we will obviously not restart the whole app in between the speeches.