Add key press animation
Now the cooldown of the button will be visually represented with a push animation. Changing the cooldown time will speed up or slow down the animation accordingly. Too low cooldown times seem to break it for some reason, 0.5 second is totally enough and works perfectly as far as I can tell.